Clarify some path variable declarations
Path is being defined in at least three places in the following order:

1. `/etc/zprofile` calls out to `/usr/libexec/path_helper`
2. Global modifications are added in `.zshrc`
3. Custom path alterations can be added in `.localrc`

In step 1 above, `/etc/paths` now declares `/usr/local/bin` at the top
so I no longer need to worry about prepending it anymore.

Next, I am adding some general paths to the beginning of the lookup
sequence that I believe are generally applicable across machines and
re-installations.

Then, `.localrc` is executed which allows for further modification of
the path if necessary.

And finally, path is exported when all is said and done.