Setup linking for config directory
More and more applications are beginning to use the ~/.config
directory for storing config files. Add support for this setup during
dotfile installation.

I chose the specific file by file strategy in case there is a set of
deeply nested files already in place for the specific application. In
this case I may just want to pick and choose specific files to link in
while leaving the majority of the existing directory structure alone.